Move code for specimens to .testing.data module
